# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS REST API",
        "description": "Human Resource Management System API Documentation",
        "contact": {
            "email": "admin@hrms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/financial-reports/trial-balance": {
            "get": {
                "tags": [
                    "Financial Reports"
                ],
                "summary": "Get trial balance report",
                "operationId": "0077b2a8bca64623b0c91122b9cf49f9",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "with_opening",
                        "in": "query",
                        "description": "Include opening balances",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trial balance report"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/financial-reports/profit-loss": {
            "get": {
                "tags": [
                    "Financial Reports"
                ],
                "summary": "Get profit and loss statement",
                "operationId": "d7635588b1e5cb71d29d6cf798ba90e3",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profit and loss statement"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/financial-reports/balance-sheet": {
            "get": {
                "tags": [
                    "Financial Reports"
                ],
                "summary": "Get balance sheet",
                "operationId": "a79da07da203548ae523ef1d014f64d6",
                "parameters": [
                    {
                        "name": "as_of_date",
                        "in": "query",
                        "description": "Date for balance sheet",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance sheet"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/financial-reports/cash-flow": {
            "get": {
                "tags": [
                    "Financial Reports"
                ],
                "summary": "Get cash flow statement",
                "operationId": "1cc953f5949754af46731103ab32247d",
                "parameters": [
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cash flow statement"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}